consolidate constants in HttpUtil
