+++ api/src/org/labkey/api/security/AuthenticatedResponse.java
@@ -19,6 +19,7 @@ package org.labkey.api.security;
 import org.apache.commons.collections4.MultiValuedMap;
 import org.apache.commons.collections4.multimap.AbstractSetValuedMap;
 import org.apache.commons.lang3.StringUtils;
+import org.labkey.api.util.HttpUtil;
 import org.labkey.api.util.PageFlowUtil;
 
 import javax.servlet.http.HttpServletResponse;
@@ -31,6 +32,9 @@ import java.util.Locale;
 import java.util.Set;
 import java.util.TreeMap;
 import java.util.TreeSet;
+
+import static org.labkey.api.util.HttpUtil.ContentSecurityPolicyEnum.DefaultSrc;
+import static org.labkey.api.util.HttpUtil.Header.ContentSecurityPolicy;
 
 /**
  * User: matthewb
@@ -60,7 +64,7 @@ public class AuthenticatedResponse extends HttpServletResponseWrapper
     @Override
     public void addHeader(String name, String value)
     {
-        if (StringUtils.equalsIgnoreCase(name, "Content-Security-Policy"))
+        if (StringUtils.equalsIgnoreCase(name, ContentSecurityPolicy))
             throw new IllegalStateException("Use addContextSecurityPolicyHeader()");
         super.addHeader(name, value);
     }
@@ -68,80 +72,9 @@ public class AuthenticatedResponse extends HttpServletResponseWrapper
     @Override
     public void setHeader(String name, String value)
     {
-        if (StringUtils.equalsIgnoreCase(name, "Content-Security-Policy"))
+        if (StringUtils.equalsIgnoreCase(name, ContentSecurityPolicy))
             throw new IllegalStateException("Use setContextSecurityPolicyHeader()");
         super.setHeader(name, value);
-    }
-
-    public enum ContentSecurityPolicyEnum
-    {
-        ConnectSrc("connect-src"),
-        DefaultSrc("default-src"),
-        FontSrc("font-src"),
-        FrameSrc("frame-src"),
-        ImgSrc("img-src"),
-        ManifestSrc("manifest-src"),
-        ObjectSrc("object-src"),
-        PrefetchSrc("prefetch-src"),
-        ScriptSrc("script-src"),
-        ScriptSrcAttr("script-src-attr"),
-        StyleSrc("style-src"),
-        StyleSrcElem("style-src-elem"),
-        StyleSrcAttr("style-src-attr"),
-        WorkerSrc("worker-src"),
-
-        BaseURI("base-uri"),
-        Sandbox("sandbox"),
-
-        FormAction("form-action"),
-        FrameAncestors("frame-ancestors"),
-        NavigateTo("navigate-to"),
-
-        ReportUri("report-uri"),
-        ReportTo("report-to"),
-
-        RequireSriFor("require-sri-for"),
-        RequireTrustedTypesFor("require-trusted-types-for"),
-        TrustedTypes("trusted-types"),
-
-        UpgradeInsecureRequests("upgrade-insecure-requests")
-        ;
-
-        final String value;
-
-        ContentSecurityPolicyEnum(String value)
-        {
-            this.value = value;
-        }
-
-        @Override
-        public String toString()
-        {
-            return value;
-        }
-    }
-
-
-    public enum CspSourceValues
-    {
-        Self("'self'"),
-        None("'none'"),
-        UnsafeEval("'unsafe-eval'"),
-        UnsafeInline("'unsafe-inline'")
-        ;
-
-        final String value;
-
-        CspSourceValues(String value)
-        {
-            this.value = value;
-        }
-
-        @Override
-        public String toString()
-        {
-            return value;
-        }
     }
 
 
@@ -154,27 +87,27 @@ public class AuthenticatedResponse extends HttpServletResponseWrapper
         }
     };
 
-    public void addContentSecurityPolicyHeader(ContentSecurityPolicyEnum csp, CspSourceValues value)
+    public void addContentSecurityPolicyHeader(HttpUtil.ContentSecurityPolicyEnum csp, HttpUtil.CspSourceValues value)
     {
         addContentSecurityPolicyHeader(csp, value.toString());
     }
 
-    public void addContentSecurityPolicyHeader(ContentSecurityPolicyEnum csp, String... values)
+    public void addContentSecurityPolicyHeader(HttpUtil.ContentSecurityPolicyEnum csp, String... values)
     {
         // if this is a src header, copy the "defaults" from connect-src
         String target = csp.toString();
-        if (target.endsWith("-src") && !target.equals(ContentSecurityPolicyEnum.DefaultSrc.toString()) && _contentSecurityPolicy.get(target).isEmpty())
-            _contentSecurityPolicy.putAll(target, _contentSecurityPolicy.get(ContentSecurityPolicyEnum.DefaultSrc.toString()));
+        if (target.endsWith("-src") && !target.equals(DefaultSrc.toString()) && _contentSecurityPolicy.get(target).isEmpty())
+            _contentSecurityPolicy.putAll(target, _contentSecurityPolicy.get(DefaultSrc.toString()));
         _contentSecurityPolicy.putAll(target, Arrays.asList(values));
         _updateContentSecurityPolicyHeader();
     }
 
-    public void setContentSecurityPolicyHeader(ContentSecurityPolicyEnum csp, CspSourceValues value)
+    public void setContentSecurityPolicyHeader(HttpUtil.ContentSecurityPolicyEnum csp, HttpUtil.CspSourceValues value)
     {
         setContentSecurityPolicyHeader(csp, value.toString());
     }
 
-    public void setContentSecurityPolicyHeader(ContentSecurityPolicyEnum csp, String value)
+    public void setContentSecurityPolicyHeader(HttpUtil.ContentSecurityPolicyEnum csp, String value)
     {
         _contentSecurityPolicy.remove(csp.toString());
         _contentSecurityPolicy.put(csp.toString(), value);
@@ -192,7 +125,7 @@ public class AuthenticatedResponse extends HttpServletResponseWrapper
             for (String s : e.getValue())
                 content.append(" ").append(s);
         }
-        super.setHeader("Content-Security-Policy", content.toString());
+        super.setHeader(ContentSecurityPolicy, content.toString());
     }
